decomposition continued
